Ensure cert and key is created from new-cert and new-ca
It was observed that openssl can occasionally output an error
code that does not invoke the retry script if not enough system
entropy exists. This commit places a simple sanity check at the end
of the scripts returning error if a file was unable to be generated
for any reason.

+++ ansible/roles/pki/files/usr/local/bin/new-ca.sh
@@ -179,6 +179,11 @@ openssl req -config ssl.conf \
 openssl rsa -in "${TLS_FILE_PREFIX}.key" -out "${TLS_FILE_PREFIX}.key.fixed"
 mv -f "${TLS_FILE_PREFIX}.key.fixed" "${TLS_FILE_PREFIX}.key"
 
+if [[ ! -f "${TLS_FILE_PREFIX}.crt" || ! -f "${TLS_FILE_PREFIX}.key" ]]; then
+  echo "failed to output certificate and key"
+  exit 1
+fi
+
 # Copy the files to OUT_DIR
 cp -f "${TLS_FILE_PREFIX}.crt" "${TLS_FILE_PREFIX}.key" "$(abspath "${OUT_DIR}")"
